requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5
openai>=1.30.0
anthropic>=0.25.0
python-dotenv>=1.0.0
numpy>=1.26.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
